If only one conformer is available, spawn the requested job
When directly specifying xyz in an ARCSpecies object, the conformer list
is updated. If the user did not ask for conformers to be spawned, ARC
will skip conformer generation. This commit allow ARC to directly
execute the requested job (e.g., freq / sp), because sometimes the user
asks not to spawn opt at all.