Minor: remove default argument of ess_trsh_methods
The default argument was None. However, this is not needed because a value is always passed in when the function is called in scheduler.py. On the off chance that the default value of None was used, it would throw an error in ~line 444 in trshTest.py since an iterable (like a list) was expected