Add same pivot check and sublist check
Add is_same_pivot to common.py for a handy function to check if two torsions share the same pivots. Add is_same_sequence_sublist to figure out case like if [1,2,3] in [1,2,3,4] (True) or [2,1,3,4] (False) or [6,1,2,3,4] (True). Helpful in scan troubleshooting to know which torsions have only one atom different. Can be helpful to know if sequence are retained in a list.