Allow more flexible dihedral reset when lower conformer found
This is related to the change in the check_scan_quality. If a conformer-based method is used. we can obtain the ICs of the lower conformer and use these values to update the new guess. Some of the dihedral change related to dihedrals not in the rotor list, thus add an option in set_dihedral to allow more flexibility.