BugFix: almost_equal_coord_lists, only check the first element
Although the function is designed to find if there are at least one common xyzs, in most cases, it actually just compared the first entry in the `xyz1`: If the first entries in both lists are the same, then return True, works okay; else if an entry in the `xyz2` has a different xyz symbols, and it is ahead of the entry equal to the first entry in `xyz1`, this will return False, not okay; else if no entry in the `xyz2` equal to the first entry in `xyz1`, it will stop iterate other entry in `xyz1` and return False, not okay. The new algorithm return `True` immediately finding a match.