Add a decorator to check the connections
By using a decorator, we can check the ssh and sftp connections before running the function. If the connection is invalid, then the decorator will also try to reconnect, and only execute the function when successfully connected. For efficiency, since the connection will be checked when calling `_send_command_to_server()`, those methods which are based on this method need not to be double checked.