BugFix: correctly spawn scan jobs
Previously, there are several bugs: 1. If there is a non-torsion, not all scan jobs are spawned due to the use of `return`. We changed it to continue; 2. after the last rotor scan improvement, scan jobs are not correctly spawned in restarting or finding a lower conformer. We changed to use `success` as the criteria, since it makes more sense. The new change also provides several benefits: By modifying restart file and restart the ARC job, you can now manually fix rotor scans if you provide the path to your scan output and mark `success` as None. you can also recheck all of the scan jobs if you simply set `success` as None, which is useful when check_scan_quality() is modified.