Determine directed rotors pivots using rotors_dict directly
rather than using the number_of_rotors attribute. The number of rotors is set to be identical to the number of values in the rotors_dict attribute of a species, so this should not affect ARC's behavior, but it does affect  some of the test assertions where a rotors_dicr is defined outside the __init__ without setting the number_of_rotors attribute.