Relocated colliding_atoms() from common to species
Rewrote colliding atoms as an ARC function instead of a QCElemental wrapper.
Now we have more control over the outcome of this function and can fine-tune it
for special cases (e.g., TSs) if needed.

Since now we have to call species.vectors.calculate_distance(),
this function can no longer stay in common and was relocated to species.