Fixed SSH Private Key, Remote File Path2 and added extra status for Azure
Added the variable for the key_filename argument in the function ssh.connect() as it did not work prev.

The function get_last_modified_time() had an argument remote_file_path_2 that was optional but had no default argument set thus errored out

Azure has custom statuses, such as CF which means configuring. This has been added in to the status check so that ARC does not keep attempting to run the same job while the Nodes are configuring during first startup