Fuzzy matching basis sets
Different software may use the same basis set but have slight variation in their naming (qchem uses hyphens while gaussian doesn't). In this implementation, a csv that is user updated will be used for the fuzzy matching if the software is relevant. it will attempt to match the users provided basis on the condition they did not provide the correct basis set form for the software