Handling null bytes in additional job info
For this, Qchem provided an out.txt file that is written in a low level language and thus contains null bytes. Therefore, when originally read, using mode `r`, os.path.isfile() will error out later on in the code due to the embedding of null bytes. This implementation determines whether there are nullbytes in the out.txt. If there are, then it will use mode `rb` instead
Decoding bytes into strings

If the file read as bytes, then it will be converted to string using `utf-8`
Update reading in out/err.txt file

Checks if the file is written in low level language and appropriately opens/converts it.