Fuzzy Matching Improved
There are use cases where the user has poorly spelt the basis set and the fuzzy match matches this spelling with the wrong basis set. We do not want this to happen. The cut off is 99 score and if there are more than 1 matches above 99, then the function will return an error to the user.

However, if there are no matches, then a hyphen will be added to the user input basis set at different positions to find a closer match. If found, then sets the basis to that. If not, then error is raised that the basis set is not supported.