Modified the job checking through SSH
Previously there were a lot of code that wasn't used due to the for and if loop at the start. They have been removed an now it will check the status correctly. Furthermore, SLURM SSH will now check the nodes to see how long they are taking to configure. If over 15 mins, then it will do an scontrol show node to see if it is not responding and return errored