Trsh Gaussian Combinations
Currently, when we troubleshoot Gaussian jobs, we don't combine troubleshoots. For example, if we had to remove the chkfile for trouleshooting but then it failed again because it required 'scf=(qc,nosymm)', then it would troubleshoot with the scf argument but 'forget' about the removal of the chkfile. This fix up allows combinations of troubleshoot methods