Don't consider conformers for which RDKit could not compute energy
Since we don't know how to rank the conformers by energy. This could potentially be problematic if one of these conformers has a lower E, but we can't return too much confs to the scheduler to be DFT'ed. We think that ARC's DFT scans may compensate if indeed a lower conformer exists.
In addition, if all conformers have no FF energies, we return them all (a hypothetical case).