ARC Job Trash Update - Trsh Job Queue
A troubleshooting function when queue walltime is reached. This will attempt first check if the user has provided queues that can be used. If not, it will then query the server (assuming PBS) to see if there are any queues that are available to user and have higher walltime

style changes