Reconstruct the flow for multi_species opt job
After the job is done, we now create {label}_energy_geo_summary.yml, which includes info of energy and geometry of the optimized cluster species.

Skip species that can not converge due to exceed step number

Parse the geometry of the successful one.
Continue with the ones after the unconverged ones.