Tests: Removed test symmetry from species tests
Symmetry was determined by Arkane, it requires the QM module in RMG, which we don't want to implement. It is actually not used in ARC anyway, it was previously passes to Arkane. So now Arkane will compute it when it's called in a sub process, no need to test it here.