Fixes orca nprocs setting
Addresses an issue where the `orca_nprocs` setting was not correctly being applied, limiting the number of processors used by orca.

This change ensures that the user-defined `orca_nprocs` value in the settings is respected, or defaults to 1, correctly determining the CPU cores to use. Also updates the default `orca_nprocs` in settings to 1.